+++ sys/mips/beri/files.beri
@@ -42,4 +42,5 @@ mips/cheri/cheri_stack.c		optional cpu_cheri
 mips/cheri/cheri_syscall.c		optional cpu_cheri
 mips/cheri/cheri_test.c			optional cpu_cheri
 mips/cheri/cheriabi_machdep.c		optional cpu_cheri compat_cheriabi
+mips/cheri/hybridabi_machdep.c		optional cpu_cheri
 mips/mips/tick.c			standard
